requirements.txt
streamlit
requests
python-dotenv
chromadb
pandas
pyyaml
